Move login and logout methods to UserController
